refactor: Collocate tests
