+++ src/components/Alert/index.test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { render, fireEvent } from '@testing-library/react'
 import { matchers } from 'jest-emotion'
-import { Alert } from '../src/components/Alert'
-import { color } from '../src/theme'
+import { Alert } from './'
+import { color } from '../../theme'
 
 expect.extend(matchers)
