+++ src/components/Avatar/index.test.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render } from '@testing-library/react'
-import { Avatar } from '../src/components/Avatar'
+import { Avatar } from './'
 
 describe('Avatar', () => {
   it('renders without crashing', () => {
