+++ src/components/Toast/index.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render, fireEvent } from '@testing-library/react'
-import { Button } from '../src/components/Button'
-import { ToastProvider, ToastConsumer, Toast } from '../src/components/Toast'
+import { Button } from '../Button'
+import { ToastProvider, ToastConsumer, Toast } from './'
 
 describe('Toast', () => {
   it('shows button and renders toasts on button-click', () => {
