+++ src/utils/callAll/index.js
@@ -1,0 +1,2 @@
+export const callAll = (...fns) => (...args) =>
+  fns.forEach(fn => fn && fn(...args))
