+++ src/utils/clamp/index.js
@@ -1,0 +1,12 @@
+export const clamp = (value, min, max) =>
+  min < max
+    ? value < min
+      ? min
+      : value > max
+      ? max
+      : value
+    : value < max
+    ? max
+    : value > min
+    ? min
+    : value
