+++ src/utils/isServer/index.js
@@ -1,0 +1,3 @@
+export function isServer() {
+  return !(typeof window !== 'undefined' && window.document)
+}
