+++ src/utils/stopPropagation/index.js
@@ -1,0 +1,1 @@
+export const stopPropagation = e => e.stopPropagation()
