+++ test/jest-common.js
@@ -5,7 +5,7 @@ const path = require('path')
 module.exports = {
   rootDir: path.join(__dirname, '..'),
   testPathIgnorePatterns: ['/node_modules/', '/coverage/', '/docs/', '/dist/'],
-  coveragePathIgnorePatterns: ['/node_modules/', '/__tests__/'],
+  coveragePathIgnorePatterns: ['/node_modules/'],
   watchPlugins: [
     'jest-watch-typeahead/filename',
     'jest-watch-typeahead/testname',
